Added the progress bar
Haven't implemented how the progress bar will increase. The idea is that this class will read from a file filled with True and False; these True and False values will represent the amount of reading and quiz the users have done and will store it on a seperate file. Using the formula (amount(Trues)/amount(all values) * 100), you can calculate the percentage and set the Progress Bar. Whenever the User returns to the subject's homepage, it will update the progress bar everytime.